refactor(browser): remove legacy Gemini Playwright script module
- move selector imports to the authoritative browser adapter scripts
- remove deprecated gemini_playwright_scripts module
- eliminate duplicated Playwright DOM selector definitions
- consolidate Gemini browser automation assets under browser adapters
